Get Subordinate Administrative Regions from location

Get Subordinate Administrative Regions from location

getAdmin(
  keywords,
  key = NULL,
  subdistrict = NULL,
  page = NULL,
  offset = NULL,
  extensions = NULL,
  filter = NULL,
  callback = NULL,
  output = "data.table",
  keep_bad_request = TRUE,
  ...
)

Arguments

keywords
Required.
Search keywords.
Rules: Country/Region, Province/State, City, County/District, Town, Country, Road, Number, Room, Building.

key
Optional.
Amap Key.
Applied from 'AutoNavi' Map API official websitehttps://lbs.amap.com/dev/

subdistrict
Optional.
Subordinate Administrative Level.
Display level of subordinate administrative regions. Available value: 0,1,2,3.
`0` do not return subordinate administrative regions.
`1` return first one subordinate administrative regions.
`2` return first two subordinate administrative regions.
`3` return first three subordinate administrative regions.

page
Optional.
Which page to return.
Each time the outmost layer will return a maximum of 20 records. If the limit is exceeded, please request the next page of records with the page argument.

offset
Optional.
Maximum records per page.
Maximum value is 20.

extensions
Optional.
Return results controller.
`base`: does not return the coordinates of the administrative district boundary.
`all`: returns only the boundary value of the current query district, not the boundary value of the child node.

filter
Optional.
Filter administrative regions.
Filtering by designated administrative divisions, which returns information only for the province/municipality.
It is strongly recommended to fill in this parameter in order to ensure the correct records.

callback
Optional.
Callback Function.
The value of callback is the customized function. Only available with JSON output. If you don't understand, it means you don't need it, just like me.

output
Optional.
Output Data Structure.
Support JSON, XML and data.table. The default value is data.table.

keep_bad_request
Optional.
Keep Bad Request to avoid breaking a workflow, especially meaningful in a batch request

max_core
Deprecated.
Since 0.5.1, parallel operation has been replaced by `furrr::future_map`, and this argument does not work anymore
A threshold of max cores for parallel operation. There is no need to set a `max_core` generally. But for some extreme high performance case, like `AMD Threadripper` and `Intel Xeon`, super multiple-core CPU will meet the limitation of queries per second.

to_table
Deprecated.
Transform response content to data.table. Since 0.5.1, this argument was merged into `output`.

Value

Returns a JSON or XML of results containing detailed subordinate administrative region information. See https://lbs.amap.com/api/webservice/guide/api/district for more information.

See also

Examples

if (FALSE) {
library(amapGeocode)

# Before the `getAdmin()` is executed,
# the token should be set by `option(amap_key = 'key')`
# or set by key argument in `getAdmin()`

# Get subordinate administrative regions as a data.table
getAdmin("Sichuan Province")
# Get subordinate administrative regions as a XML
getCoord("Sichuan Province", output = "XML")
}